requirements.txt
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
Django==5.2
django-picklefield==3.3
django-q==1.3.9
django-qr==0.2.0
django-redis==5.4.0
djangorestframework==3.16.0
Faker==37.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==3.5.3
segno==1.6.6
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
wcwidth==0.2.13
